+++ sundog/src/cluster/absence.rs
@@ -103,6 +103,7 @@ mod tests {
             gossip_addr: SocketAddr::from((Ipv4Addr::LOCALHOST, 7000)),
             data_addr: SocketAddr::from((Ipv4Addr::LOCALHOST, 8000)),
             incarnation: 1,
+            protocol: crate::wire::PROTOCOL_VERSION,
         }
     }